Begin renaming things
It is better to call things traveler as that is actually what is being
run.
