+++ files/traveler/find-sequences.sql
